Langgraph_multiagent_travel/backend/requirements.txt
(# Core web / runtime)
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.7
python-dotenv>=1.0.0
requests>=2.28.0

(# LangChain / LLM integrations - use compatible versions for your environment)
langchain>=0.0.280
langchain-core>=0.0.200
langchain-community>=0.2.0

# Optional/adapters (uncomment or install as needed
langchain-groq
langchain-openai
langchain-tavily
langchain-google-community
langchain-google-community[places]

(# Langgraph - core project library)
(# Utilities used in code)
python-dotenv>=1.0.0

